… CHANGELOG.md
… OpenSim/Common/SmoothSegmentedFunction.h