This PR de-duplicates the data held in `SmoothSegmentedFunction` by implementing a global cache for its data, such that when the cache is given identical curve parameters, it returns a read-only, shared copy of the curve data.

The significance of de-duplicating `SmoothSegmentedFunction` is that it's used extensively by muscle components such as `Millard2012EquilibriumMuscle`. The Millard muscle uses 5 different curves, but between individual muscle instances these curves are typically identical. Taking the `RajagopalModel`  as example:

We find the following `SmoothSegmentedFunction` curves in this model, each holding multiple `SimTK::Spline`:
- `ActiveForceLengthCurve` -> 5 splines
- `ForceVelocityInverseCurve` -> 4 splines
- `FiberForceLengthCurve` -> 2 splines
- `TendonForceLengthCurve` -> 2 splines
- `ForceVelocityCurve` -> 4 splines

With each spline ~ 200 doubles = 1.6 kB
Before deduplication the Rajagopal model has ~80 copies of each of the above curves which sums up to ~2MB. Applying deduplication reduces this to ~27kB.

### Brief summary of changes
The changes apply to `SmoothSegmentedFunction.h/cpp`:
- `SmoothSegmentedFunctionParameters` is added to collect all parameters that are used to construct a new smooth function, e.g. `_x0`, `_x1` etc...
- `SmoothSegmentedFunctionData` is added to collect all data required to evaluate the smooth function, e.g. `_arraySplineUX_`, etc. This data can be constructed from the `SmoothSegmentedFunctionParameters`.
- `SmoothSegmentedFunctionDataCache` is added to hold an `unordered_map` of smart pointers to the data, with the hashed-parameters used as key. Whenever a muscle curve is constructed, the parameters are used to obtain a pointer to already existing data, or to construct a new one.
- `SmoothSegmentedFunction` holds a shared pointer to the `SmoothSegmentedFunctionData`.

### Performance Results

Performance was evaluated against current master:

| Test Name | lhs [secs] | σ [secs] | rhs [secs] | σ [secs] | Speedup |
| -------------------------- | ---------- | -------- | ---------- | -------- | ------- |
|`ToyDropLanding_nomuscles` | 0.03 | 0.00 | 0.03 | 0.00 | 1.10 | |
|`passive_dynamic` | 1.79 | 0.00 | 1.79 | 0.00 | 1.00 |
| `RajagopalModel` | 4.66 | 0.00 | 4.35 | 0.01 | 1.07 | 
| `Arm26` | 0.11 | 0.00 | 0.10 | 0.00 | 1.02 | 
| `Gait2354` | 0.14 | 0.00 | 0.14 | 0.00 | 1.02 | 
| `ToyDropLanding` | 4.74 | 0.00 | 4.70 | 0.00 | 1.01 | 
| `passive_dynamic_noanalysis` | 1.17 | 0.00 | 1.18 | 0.00 | 1.00 |

As expected, the `RajagopalModel` can be seen to have the most significant performance boost, because it has many `Millard2012EquilibriumMuscle`s.

### CHANGELOG.md

- updated.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/opensim-org/opensim-core/3442)
<!-- Reviewable:end -->