@pepbos 2023-05-15
> LGTM! Seems like Adam very thoroughly reviewed this already.
> 
> Cool changes. I learned some stuff about smart pointer management.
> 
> > Reviewed 1 of 2 files at r2, 2 of 2 files at r6, all commit messages.
> > _[Reviewable](https://reviewable.io/reviews/opensim-org/opensim-core/3442#-:-NVGHUCJ0SncomUhoLLc:b-f80pnr)_ status: all files reviewed, 35 unresolved discussions (waiting on @adamkewley and @pepbos)

Great! Thanks for reviewing! I hope the CI will pass now.